Implement the BFR validation
